+++ lib/active_force/bulk/records.rb
@@ -3,6 +3,8 @@ require 'csv'
 module ActiveForce
   module Bulk
     class Records
+      NULL_VALUE = '#N/A'.freeze
+
       attr_reader :headers, :data
       def initialize(headers:, data:)
         @headers = headers
@@ -16,9 +18,24 @@ module ActiveForce
       end
 
       def self.parse_from_attributes(records)
+        # Sorting ensures that the headers line up with the values for the CSV
         headers = records.first.keys.sort.map(&:to_s)
-        data = records.map { |r| r.sort.pluck(-1) }
+        data = records.map do |r|
+           r.transform_values { |v| transform_value_for_sf(v) }.sort.pluck(-1)
+        end
         new(headers: headers, data: data)
+      end
+
+      # SF expects a special value for setting a column to be NULL.
+      def self.transform_value_for_sf(value)
+        case value
+        when NilClass
+          NULL_VALUE
+        when Time
+          value.iso8601
+        else
+          value.to_s
+        end
       end
     end
   end